+++ src/main/java/org/truffleruby/core/time/RubyTimeOutputFormatter.java
@@ -46,6 +46,13 @@ public class RubyTimeOutputFormatter {
         if (flags.indexOf('-') != -1) { // no padding
             return 0;
         }
+        return getWidthAlwaysPad(defaultWidth);
+    }
+
+
+    // For non-numerical output flag "-" should be ignored.
+    // On the other hand in fact it's applied to some non-numerical directives, e.g. %b, %B etc.
+    public int getWidthAlwaysPad(int defaultWidth) {
         return this.width != 0 ? this.width : defaultWidth;
     }
 
@@ -61,6 +68,22 @@ public class RubyTimeOutputFormatter {
                     break;
                 case '-': // no padding
                     padder = '\0';
+                    break;
+            }
+        }
+        return padder;
+    }
+
+    // some directives treat "-" in some different way, e.g. %z
+    public char getPadderAlwaysPad(char defaultPadder) {
+        char padder = defaultPadder;
+        for (int i = 0; i < flags.length(); i++) {
+            switch (flags.charAt(i)) {
+                case '_':
+                    padder = ' ';
+                    break;
+                case '0':
+                    padder = '0';
                     break;
             }
         }
@@ -118,6 +141,10 @@ public class RubyTimeOutputFormatter {
                 return padding(Long.toString(value), width, padder);
             }
         }
+    }
+
+    public boolean hasNoPaddingFlag() {
+        return flags.contains("-");
     }
 
     private static final int SMALLBUF = 100;