+++ src/main/java/org/truffleruby/core/time/TimeNodes.java
@@ -13,7 +13,9 @@ import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.dsl.Cached;
 import com.oracle.truffle.api.dsl.Specialization;
+import com.oracle.truffle.api.library.CachedLibrary;
 import com.oracle.truffle.api.object.Shape;
+import com.oracle.truffle.api.profiles.BranchProfile;
 import com.oracle.truffle.api.profiles.ConditionProfile;
 import com.oracle.truffle.api.strings.AbstractTruffleString;
 import com.oracle.truffle.api.strings.TruffleString;
@@ -36,6 +38,7 @@ import org.truffleruby.core.time.RubyDateFormatter.Token;
 import org.truffleruby.language.Nil;
 import org.truffleruby.annotations.Visibility;
 import org.truffleruby.language.control.RaiseException;
+import org.truffleruby.language.library.RubyLibrary;
 import org.truffleruby.language.library.RubyStringLibrary;
 import org.truffleruby.language.objects.AllocationTracing;
 
@@ -127,6 +130,26 @@ public abstract class TimeNodes {
 
     }
 
+    @Primitive(name = "time_utctime")
+    public abstract static class UtcTimeNode extends PrimitiveArrayArgumentsNode {
+
+        @Specialization
+        protected RubyTime utc(RubyTime time) {
+            time.isUtc = true;
+            time.relativeOffset = false;
+            time.zone = coreStrings().UTC.createInstance(getContext());
+            time.dateTime = inUTC(time.dateTime);
+
+            return time;
+        }
+
+        @TruffleBoundary
+        private ZonedDateTime inUTC(ZonedDateTime dateTime) {
+            return dateTime.withZoneSameInstant(GetTimeZoneNode.UTC);
+        }
+    }
+
+
     @Primitive(name = "time_add")
     public abstract static class TimeAddNode extends PrimitiveArrayArgumentsNode {
         @TruffleBoundary
@@ -141,8 +164,22 @@ public abstract class TimeNodes {
     @CoreMethod(names = { "gmtime", "utc" })
     public abstract static class GmTimeNode extends CoreMethodArrayArgumentsNode {
 
-        @Specialization
-        protected RubyTime gmtime(RubyTime time) {
+        @Specialization(limit = "getRubyLibraryCacheLimit()")
+        protected RubyTime gmtime(RubyTime time,
+                @Cached BranchProfile errorProfile,
+                @Cached BranchProfile notModifiedProfile,
+                @CachedLibrary("time") RubyLibrary rubyLibrary) {
+
+            if (time.isUtc) {
+                notModifiedProfile.enter();
+                return time;
+            }
+
+            if (rubyLibrary.isFrozen(time)) {
+                errorProfile.enter();
+                throw new RaiseException(getContext(), coreExceptions().frozenError(time, this));
+            }
+
             final ZonedDateTime dateTime = time.dateTime;
 
             time.isUtc = true;
@@ -437,6 +474,7 @@ public abstract class TimeNodes {
                     pattern,
                     time.dateTime,
                     time.zone,
+                    time.isUtc,
                     getContext(),
                     getLanguage(),
                     this,
@@ -498,7 +536,8 @@ public abstract class TimeNodes {
         private RubyTime buildTime(RubyLanguage language, RubyClass timeClass, int sec, int min, int hour, int mday,
                 int month,
                 int year, int nsec, int isdst, boolean isutc, Object utcoffset) {
-            if (sec < 0 || sec > 60 || // MRI accepts sec=60, whether it is a leap second or not
+            if (nsec < 0 || nsec > 999999999 ||
+                    sec < 0 || sec > 60 || // MRI accepts sec=60, whether it is a leap second or not
                     min < 0 || min > 59 ||
                     hour < 0 || hour > 23 ||
                     mday < 1 || mday > 31 ||